HBASE-25793 Move BaseLoadBalancer.Cluster to a separated file (#3185)
