+++ hbase-balancer/src/main/java/org/apache/hadoop/hbase/master/balancer/BalanceAction.java
@@ -1,0 +1,55 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.hadoop.hbase.master.balancer;
+
+import org.apache.yetus.audience.InterfaceAudience;
+
+/**
+ * An action to move or swap a region
+ */
+@InterfaceAudience.Private
+abstract class BalanceAction {
+  enum Type {
+    ASSIGN_REGION, MOVE_REGION, SWAP_REGIONS, NULL,
+  }
+
+  static final BalanceAction NULL_ACTION = new BalanceAction(Type.NULL) {
+  };
+
+  private final Type type;
+
+  BalanceAction(Type type) {
+    this.type = type;
+  }
+
+  /**
+   * Returns an Action which would undo this action
+   */
+  BalanceAction undoAction() {
+    return this;
+  }
+
+  Type getType() {
+    return type;
+  }
+
+  @Override
+  public String toString() {
+    return type + ":";
+  }
+}
