+++ hbase-server/src/main/java/org/apache/hadoop/hbase/master/balancer/CandidateGenerator.java
@@ -28,7 +28,7 @@ import org.apache.yetus.audience.InterfaceAudience;
 @InterfaceAudience.Private
 abstract class CandidateGenerator {
 
-  abstract BaseLoadBalancer.Cluster.Action generate(BaseLoadBalancer.Cluster cluster);
+  abstract BalanceAction generate(BalancerClusterState cluster);
 
   /**
    * From a list of regions pick a random one. Null can be returned which
@@ -42,7 +42,7 @@ abstract class CandidateGenerator {
    * @return a random {@link RegionInfo} or null if an asymmetrical move is
    *   suggested.
    */
-  int pickRandomRegion(BaseLoadBalancer.Cluster cluster, int server,
+  int pickRandomRegion(BalancerClusterState cluster, int server,
     double chanceOfNoSwap) {
     // Check to see if this is just a move.
     if (cluster.regionsPerServer[server].length == 0
@@ -54,7 +54,7 @@ abstract class CandidateGenerator {
     return cluster.regionsPerServer[server][rand];
   }
 
-  int pickRandomServer(BaseLoadBalancer.Cluster cluster) {
+  int pickRandomServer(BalancerClusterState cluster) {
     if (cluster.numServers < 1) {
       return -1;
     }
@@ -62,7 +62,7 @@ abstract class CandidateGenerator {
     return StochasticLoadBalancer.RANDOM.nextInt(cluster.numServers);
   }
 
-  int pickRandomRack(BaseLoadBalancer.Cluster cluster) {
+  int pickRandomRack(BalancerClusterState cluster) {
     if (cluster.numRacks < 1) {
       return -1;
     }
@@ -70,7 +70,7 @@ abstract class CandidateGenerator {
     return StochasticLoadBalancer.RANDOM.nextInt(cluster.numRacks);
   }
 
-  int pickOtherRandomServer(BaseLoadBalancer.Cluster cluster, int serverIndex) {
+  int pickOtherRandomServer(BalancerClusterState cluster, int serverIndex) {
     if (cluster.numServers < 2) {
       return -1;
     }
@@ -82,7 +82,7 @@ abstract class CandidateGenerator {
     }
   }
 
-  int pickOtherRandomRack(BaseLoadBalancer.Cluster cluster, int rackIndex) {
+  int pickOtherRandomRack(BalancerClusterState cluster, int rackIndex) {
     if (cluster.numRacks < 2) {
       return -1;
     }
@@ -94,10 +94,10 @@ abstract class CandidateGenerator {
     }
   }
 
-  BaseLoadBalancer.Cluster.Action pickRandomRegions(BaseLoadBalancer.Cluster cluster,
+  BalanceAction pickRandomRegions(BalancerClusterState cluster,
     int thisServer, int otherServer) {
     if (thisServer < 0 || otherServer < 0) {
-      return BaseLoadBalancer.Cluster.NullAction;
+      return BalanceAction.NULL_ACTION;
     }
 
     // Decide who is most likely to need another region
@@ -114,20 +114,20 @@ abstract class CandidateGenerator {
     return getAction(thisServer, thisRegion, otherServer, otherRegion);
   }
 
-  protected BaseLoadBalancer.Cluster.Action getAction(int fromServer, int fromRegion,
+  protected BalanceAction getAction(int fromServer, int fromRegion,
       int toServer, int toRegion) {
     if (fromServer < 0 || toServer < 0) {
-      return BaseLoadBalancer.Cluster.NullAction;
+      return BalanceAction.NULL_ACTION;
     }
     if (fromRegion >= 0 && toRegion >= 0) {
-      return new BaseLoadBalancer.Cluster.SwapRegionsAction(fromServer, fromRegion,
+      return new SwapRegionsAction(fromServer, fromRegion,
         toServer, toRegion);
     } else if (fromRegion >= 0) {
-      return new BaseLoadBalancer.Cluster.MoveRegionAction(fromRegion, fromServer, toServer);
+      return new MoveRegionAction(fromRegion, fromServer, toServer);
     } else if (toRegion >= 0) {
-      return new BaseLoadBalancer.Cluster.MoveRegionAction(toRegion, toServer, fromServer);
+      return new MoveRegionAction(toRegion, toServer, fromServer);
     } else {
-      return BaseLoadBalancer.Cluster.NullAction;
+      return BalanceAction.NULL_ACTION;
     }
   }
 }
